refac: refactor dropzone overly and files store
